requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
filelock==3.13.1
frozenlist==1.4.1
graphql-core==3.2.3
idna==3.6
multidict==6.0.5
numpy==1.26.4
pandas==2.2.1
psutil==5.9.8
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
scipy==1.12.0
sgqlc==16.3
shinny-structlog==0.0.4
simplejson==3.19.2
six==1.16.0
tqsdk==3.5.3
tqsdk-ctpse==1.0.2
tqsdk-sm==1.0.5
tzdata==2024.1
urllib3==2.2.1
websockets==12.0
yarl==1.9.4
